+++ packages/components/src/Menu/MenuItem.test.tsx
@@ -33,9 +33,11 @@ import { MenuItem } from './MenuItem'
 test('MenuItem', () => {
   assertSnapshot(<MenuItem>who!</MenuItem>)
 })
+
 test('MenuItem - icon', () => {
   assertSnapshot(<MenuItem icon="Beaker">who!</MenuItem>)
 })
+
 test('MenuItem - detail', () => {
   assertSnapshot(<MenuItem detail="Is an excellent question">who!</MenuItem>)
 })
@@ -44,6 +46,23 @@ test('MenuItem - current', () => {
   assertSnapshot(
     <MenuItem current icon="Home">
       who!
+    </MenuItem>
+  )
+})
+
+test('MenuItem - artwork', () => {
+  assertSnapshot(
+    <MenuItem
+      artwork={
+        <svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path
+            d="M12 2C6.48 2 2 6.48 2 12C2 17.52 6.48 22 12 22C17.52 22 22 17.52 22 12C22 6.48 17.52 2 12 2ZM13 17H11V15H13V17ZM13 13H11V7H13V13Z"
+            fill="#7FFFD4"
+          />
+        </svg>
+      }
+    >
+      Artwork
     </MenuItem>
   )
 })